Remove card trial from onboarding
