+++ src-tauri/src/lib.rs
@@ -241,9 +241,6 @@ pub fn run() {
             os_accounts::os_accounts_top_up,
             os_accounts::os_accounts_open_portal,
             os_accounts::os_accounts_referral_summary,
-            os_accounts::os_accounts_prepare_trial_checkout,
-            os_accounts::os_accounts_start_trial_checkout,
-            focus_main_window
         ])
         .manage(RecordingPresenceBoundsState::default())
         .manage(hermes_bridge::HermesBridge::default())
@@ -453,21 +450,6 @@ fn repair_agent_task_statuses_on_app_start(app: &tauri::App) {
             ),
         }
     });
-}
-
-/// Bring the app back to the foreground. The trial flow calls this when the
-/// subscription poll detects checkout finished in the browser, so the user
-/// lands back in June without hunting for the window.
-#[tauri::command]
-fn focus_main_window(app: tauri::AppHandle) {
-    #[cfg(target_os = "macos")]
-    show_main_window(&app);
-    #[cfg(not(target_os = "macos"))]
-    if let Some(window) = app.get_webview_window("main") {
-        let _ = window.show();
-        let _ = window.unminimize();
-        let _ = window.set_focus();
-    }
 }
 
 fn close_main_window(app: &tauri::AppHandle) {
