+++ src/test/onboarding.test.tsx
@@ -25,10 +25,7 @@ const mocks = vi.hoisted(() => ({
   osAccountsLogin: vi.fn(),
   scribeOpenVerifyPage: vi.fn(),
   osAccountsCancelLogin: vi.fn(),
-  osAccountsPrepareTrialCheckout: vi.fn(),
-  osAccountsStartTrialCheckout: vi.fn(),
   osAccountsOpenPortal: vi.fn(),
-  focusMainWindow: vi.fn(),
   listen: vi.fn(),
 }));
 
@@ -42,23 +39,18 @@ vi.mock("../lib/tauri", () => ({
   osAccountsLogin: mocks.osAccountsLogin,
   scribeOpenVerifyPage: mocks.scribeOpenVerifyPage,
   osAccountsCancelLogin: mocks.osAccountsCancelLogin,
-  osAccountsPrepareTrialCheckout: mocks.osAccountsPrepareTrialCheckout,
-  osAccountsStartTrialCheckout: mocks.osAccountsStartTrialCheckout,
   osAccountsOpenPortal: mocks.osAccountsOpenPortal,
-  focusMainWindow: mocks.focusMainWindow,
 }));
 
 vi.mock("@tauri-apps/api/event", () => ({
   listen: mocks.listen,
 }));
 
-// Signed in AND already on a subscription: the trial step auto-skips, so the
-// full-walk test exercises the same path an existing member re-running the
-// wizard sees.
 const account: AccountStatus = {
   signedIn: true,
   configured: true,
   user: { id: "u1", handle: "casey", displayName: "Casey Tester" },
+  balance: { credits: 5000, usdMillis: 5000 },
   subscription: { subscribed: true, status: "trialing" },
 };
 
@@ -120,19 +112,14 @@ function shortcut(label: string) {
 
 describe("OnboardingFlow", () => {
   let emitDictationEvent: ListenHandler | undefined;
-  let emitBillingCallback: ListenHandler | undefined;
 
   beforeEach(() => {
     vi.clearAllMocks();
     localStorage.clear();
     emitDictationEvent = undefined;
-    emitBillingCallback = undefined;
     mocks.listen.mockImplementation(
       (eventName: string, handler: ListenHandler) => {
         if (eventName === "dictation-event") emitDictationEvent = handler;
-        if (eventName === "os-accounts-billing-callback") {
-          emitBillingCallback = handler;
-        }
         return Promise.resolve(vi.fn());
       },
     );
@@ -142,11 +129,7 @@ describe("OnboardingFlow", () => {
     );
     mocks.openPrivacySettings.mockResolvedValue(undefined);
     mocks.osAccountsCancelLogin.mockResolvedValue(undefined);
-    mocks.osAccountsPrepareTrialCheckout.mockResolvedValue({
-      outcome: "ready",
-    });
     mocks.osAccountsOpenPortal.mockResolvedValue(undefined);
-    mocks.focusMainWindow.mockResolvedValue(undefined);
     mocks.setDictationLanguage.mockResolvedValue(undefined);
     mocks.setDictationShortcut.mockResolvedValue(undefined);
     mocks.dictationSettings.mockResolvedValue({
@@ -166,7 +149,6 @@ describe("OnboardingFlow", () => {
     return {
       account,
       onAccountChanged: vi.fn(),
-      onRefreshAccount: vi.fn(async () => undefined),
       onComplete: vi.fn(),
       ...overrides,
     };
@@ -233,8 +215,8 @@ describe("OnboardingFlow", () => {
       expect(screen.getByRole("button", { name: "Continue" })).toBeEnabled(),
     );
     await user.click(screen.getByRole("button", { name: "Continue" }));
-    // The trial auto-skips (already subscribed), landing on the practice
-    // step. Typing into the field stands in for dictation.
+    // The next step is hands-on practice. Onboarding no longer opens billing
+    // or asks for a card before the user tries the product.
     const input = await screen.findByPlaceholderText(/Tell June what to do/i);
     await user.type(input, "hello there");
     await screen.findByRole("status", { name: "Dictation is working" });
@@ -390,16 +372,6 @@ describe("OnboardingFlow", () => {
     );
   });
 
-  async function walkToTrial(user: ReturnType<typeof userEvent.setup>) {
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-    grantPermissions();
-    await waitFor(() =>
-      expect(screen.getByRole("button", { name: "Continue" })).toBeEnabled(),
-    );
-    await user.click(screen.getByRole("button", { name: "Continue" }));
-    await screen.findByRole("heading", { name: "Start your free trial" });
-  }
-
   it("signs the user in from the first step", async () => {
     const user = userEvent.setup();
     const onAccountChanged = vi.fn();
@@ -456,281 +428,22 @@ describe("OnboardingFlow", () => {
     }
   });
 
-  it("starts the trial checkout in one click and advances when the subscription lands", async () => {
+  it("does not ask unsubscribed users for a card during onboarding", async () => {
     const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockResolvedValue({
-      outcome: "checkoutOpened",
-    });
-    const props = flowProps({ account: unsubscribedAccount });
-    const { rerender } = render(<OnboardingFlow {...props} />);
+    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
     await screen.findByRole("heading", { name: "Let June listen and type" });
 
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-
-    expect(mocks.osAccountsStartTrialCheckout).toHaveBeenCalledOnce();
-    // No portal page in the middle: the direct checkout opened, so the
-    // portal command must not have fired.
-    expect(mocks.osAccountsOpenPortal).not.toHaveBeenCalled();
-    await screen.findByText(/Waiting for trial/);
-    await screen.findByText(/Finish in Stripe checkout/);
-
-    // Checkout completes in the browser; the refreshed snapshot flips the
-    // step to its success state and pulls the app forward.
-    rerender(<OnboardingFlow {...props} account={account} />);
-    await screen.findByRole("heading", {
-      name: "You're good to go",
-    });
-    // The privacy reassurance moved from a section heading into the card
-    // subtitle; the recap below is now a labelled row of points.
-    expect(
-      screen.getByText("Your trial is live, and June keeps it all private."),
-    ).toBeInTheDocument();
-    expect(screen.getByText("Local first")).toBeInTheDocument();
-    expect(
-      screen.getByText("Your data stays on your device by default."),
-    ).toBeInTheDocument();
-    expect(screen.getByText("Private AI models")).toBeInTheDocument();
-    expect(
-      screen.getByText(
-        "Inference runs on private, zero-retention models by default.",
-      ),
-    ).toBeInTheDocument();
-    expect(screen.getByText("Minimal data retention")).toBeInTheDocument();
-    expect(
-      screen.getByText("June keeps only what's needed for your account."),
-    ).toBeInTheDocument();
-    expect(mocks.focusMainWindow).toHaveBeenCalledOnce();
-
+    grantPermissions();
+    await waitFor(() =>
+      expect(screen.getByRole("button", { name: "Continue" })).toBeEnabled(),
+    );
     await user.click(screen.getByRole("button", { name: "Continue" }));
     await screen.findByPlaceholderText(/Tell June what to do/i);
-  });
 
-  it("pre-mints the checkout session on the pitch and again after a cancel", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockResolvedValue({
-      outcome: "checkoutOpened",
-    });
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-    // The pitch isn't on screen yet, so nothing should be minted.
-    expect(mocks.osAccountsPrepareTrialCheckout).not.toHaveBeenCalled();
-
-    await walkToTrial(user);
-    await waitFor(() =>
-      expect(mocks.osAccountsPrepareTrialCheckout).toHaveBeenCalledOnce(),
-    );
-    // Prefetching is invisible: the pitch keeps its ready button.
+    expect(screen.queryByRole("heading", { name: /free trial/i })).toBeNull();
     expect(
-      screen.getByRole("button", { name: "Start free trial" }),
-    ).toBeEnabled();
-
-    // A canceled checkout consumed the prepared session; landing back on
-    // the pitch must mint a fresh one in the background.
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-    await screen.findByText(/Waiting for trial/);
-    emitBillingCallback?.({ payload: "cancel" });
-    await screen.findByRole("heading", { name: "Start your free trial" });
-    await waitFor(() =>
-      expect(mocks.osAccountsPrepareTrialCheckout).toHaveBeenCalledTimes(2),
-    );
-  });
-
-  it("keeps a failed pre-mint silent and lets the click mint on the spot", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsPrepareTrialCheckout.mockRejectedValue({
-      code: "trial_checkout_unavailable",
-      message: "Could not start the free trial checkout.",
-    });
-    mocks.osAccountsStartTrialCheckout.mockResolvedValue({
-      outcome: "checkoutOpened",
-    });
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await waitFor(() =>
-      expect(mocks.osAccountsPrepareTrialCheckout).toHaveBeenCalled(),
-    );
-    // The background failure must not leak into the pitch.
-    expect(screen.queryByText(/Could not start/)).toBeNull();
-
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-    await screen.findByText(/Waiting for trial/);
-    expect(mocks.osAccountsOpenPortal).not.toHaveBeenCalled();
-  });
-
-  it("refreshes the snapshot when the pre-mint reports an existing subscription", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsPrepareTrialCheckout.mockResolvedValue({
-      outcome: "alreadySubscribed",
-    });
-    const onRefreshAccount = vi.fn(async () => account);
-    render(
-      <OnboardingFlow
-        {...flowProps({ account: unsubscribedAccount, onRefreshAccount })}
-      />,
-    );
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await waitFor(() => expect(onRefreshAccount).toHaveBeenCalled());
-  });
-
-  it("reacts to the post-checkout deep link without waiting out the poll", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockResolvedValue({
-      outcome: "checkoutOpened",
-    });
-    const onRefreshAccount = vi.fn(async () => undefined);
-    render(
-      <OnboardingFlow
-        {...flowProps({ account: unsubscribedAccount, onRefreshAccount })}
-      />,
-    );
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-    await screen.findByText(/Waiting for trial/);
-
-    // Cancel: back to the pitch with a friendly note, not an error.
-    emitBillingCallback?.({ payload: "cancel" });
-    await screen.findByRole("heading", { name: "Start your free trial" });
-    await screen.findByText(/Checkout canceled/);
-
-    // Success: the deep link triggers an immediate status refresh.
-    onRefreshAccount.mockClear();
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-    await screen.findByText(/Waiting for trial/);
-    emitBillingCallback?.({ payload: "success" });
-    await waitFor(() => expect(onRefreshAccount).toHaveBeenCalled());
-  });
-
-  it("falls back to the portal when direct checkout is unavailable", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockRejectedValue(
-      new Error("trial_checkout_unavailable"),
-    );
-    mocks.osAccountsOpenPortal.mockResolvedValue(undefined);
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-
-    await waitFor(() =>
-      expect(mocks.osAccountsOpenPortal).toHaveBeenCalledOnce(),
-    );
-    await screen.findByText(/Finish in your account portal/);
-  });
-
-  // A grant from a sign-in that predates billing:write can't mint the
-  // checkout session and refreshing can't broaden it. The hook re-runs
-  // sign-in and retries so the user still lands on Stripe, never the portal.
-  it("re-authenticates and retries when the grant lacks the billing scope", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout
-      .mockRejectedValueOnce({
-        code: "trial_checkout_needs_reauth",
-        message: "Sign in again to continue.",
-      })
-      .mockResolvedValueOnce({ outcome: "checkoutOpened" });
-    let finishLogin: (() => void) | undefined;
-    mocks.osAccountsLogin.mockImplementation(
-      () =>
-        new Promise<typeof unsubscribedAccount>((resolve) => {
-          finishLogin = () => resolve(unsubscribedAccount);
-        }),
-    );
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-
-    // While the sign-in bounce is in flight the button says what's happening
-    // (and stays disabled) instead of pretending checkout is opening.
-    const reauthButton = await screen.findByRole("button", {
-      name: "Confirming sign-in...",
-    });
-    expect(reauthButton).toBeDisabled();
-    finishLogin?.();
-
-    await waitFor(() =>
-      expect(mocks.osAccountsStartTrialCheckout).toHaveBeenCalledTimes(2),
-    );
-    expect(mocks.osAccountsLogin).toHaveBeenCalledOnce();
-    expect(mocks.osAccountsOpenPortal).not.toHaveBeenCalled();
-    await screen.findByText(/Waiting for trial/);
-    await screen.findByText(/Finish in Stripe checkout/);
-  });
-
-  it("falls back to the portal when the re-auth itself fails", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockRejectedValue({
-      code: "trial_checkout_needs_reauth",
-      message: "Sign in again to continue.",
-    });
-    mocks.osAccountsLogin.mockRejectedValue({
-      code: "network_error",
-      message: "Could not reach OS Accounts.",
-    });
-    mocks.osAccountsOpenPortal.mockResolvedValue(undefined);
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-
-    await waitFor(() =>
-      expect(mocks.osAccountsOpenPortal).toHaveBeenCalledOnce(),
-    );
-    // No retry without a fresh grant: the direct path was attempted once.
-    expect(mocks.osAccountsStartTrialCheckout).toHaveBeenCalledOnce();
-    await screen.findByText(/Finish in your account portal/);
-  });
-
-  it("falls back to the portal when re-auth does not unblock checkout", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockRejectedValue({
-      code: "trial_checkout_needs_reauth",
-      message: "Sign in again to continue.",
-    });
-    mocks.osAccountsLogin.mockResolvedValue(unsubscribedAccount);
-    mocks.osAccountsOpenPortal.mockResolvedValue(undefined);
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-
-    await waitFor(() =>
-      expect(mocks.osAccountsOpenPortal).toHaveBeenCalledOnce(),
-    );
-    expect(mocks.osAccountsStartTrialCheckout).toHaveBeenCalledTimes(2);
-    await screen.findByText(/Finish in your account portal/);
-  });
-
-  it("returns to the pitch when the user cancels the re-auth", async () => {
-    const user = userEvent.setup();
-    mocks.osAccountsStartTrialCheckout.mockRejectedValue({
-      code: "trial_checkout_needs_reauth",
-      message: "Sign in again to continue.",
-    });
-    mocks.osAccountsLogin.mockRejectedValue({
-      code: "login_canceled",
-      message: "Sign-in canceled.",
-    });
-    render(<OnboardingFlow {...flowProps({ account: unsubscribedAccount })} />);
-    await screen.findByRole("heading", { name: "Let June listen and type" });
-
-    await walkToTrial(user);
-    await user.click(screen.getByRole("button", { name: "Start free trial" }));
-
-    // Back at the pitch with a friendly note; no portal page forced open.
-    await screen.findByRole("heading", { name: "Start your free trial" });
-    await screen.findByText(/Sign-in canceled/);
+      screen.queryByRole("button", { name: /Start free trial/i }),
+    ).toBeNull();
     expect(mocks.osAccountsOpenPortal).not.toHaveBeenCalled();
   });
